+++ xlabml/stats.py
@@ -37,12 +37,12 @@ def dataset_stats(dataset: SequenceDataset, sample_size: int):
     tokenizer = token_dataset.tokenizer
     pad_index = tokenizer[tokenizer.pad_token]
     text_sample_size = min(sample_size, len(token_dataset))
-    chunk_sample_size = min(sample_size, len(dataset))
+    seq_sample_size = min(sample_size, len(dataset))
     text_sample_indices = np.random.permutation(text_sample_size)
-    chunk_sample_indices = np.random.permutation(chunk_sample_size)
+    seq_sample_indices = np.random.permutation(seq_sample_size)
     text_sample = text_dataset[text_sample_indices]
     token_sample = token_dataset[text_sample_indices]
-    chunk_sample = [dataset[int(idx)][0] for idx in chunk_sample_indices]
+    seq_sample = [dataset[int(idx)][0] for idx in seq_sample_indices]
     text_lengths = np.array([len(text) for text in text_sample])
     token_lengths = np.array([len(indices) for indices in token_sample])
     return {
@@ -54,7 +54,7 @@ def dataset_stats(dataset: SequenceDataset, sample_size: int):
         'token_length_mean': np.mean(token_lengths),
         'token_length_median': np.median(token_lengths),
         'token_length_std': np.std(token_lengths),
-        'seq_fill_ratio_mean': np.mean([(chunk != pad_index).sum() for chunk in chunk_sample]) / dataset.seq_len,
+        'seq_fill_ratio_mean': np.mean([(seq != pad_index).sum() for seq in seq_sample]) / len(dataset[0][0]),
     }